requirements.txt
streamlit
tensorflow
pillow
scikit-learn
joblib
